@@ -1,3 +1,3 @@
 module core/scanmodules
 
-go 1.21
+go 1.24.4
